refactor : mvc에 맞게 폴더구조 생성 후 해당 폴더로 이동
