+++ src/main/java/Application.java
@@ -1,5 +1,3 @@
-package lotto;
-
 import controller.LottoController;
 
 public class Application {
