+++ src/main/java/exception/CustomException.java
@@ -1,4 +1,4 @@
-package custom;
+package exception;
 
 public class CustomException extends IllegalArgumentException {
     private static final String ERROR_PREFIX = "[ERROR] ";
