+++ src/test/java/lotto/BonusNumberTest.java
@@ -2,6 +2,7 @@ package lotto;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import model.BonusNumber;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
