+++ src/test/java/lotto/WinningInfoTest.java
@@ -4,6 +4,11 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.util.List;
+import model.BonusNumber;
+import model.Lotto;
+import model.LottoRank;
+import model.WinningInfo;
+import model.WinningNumber;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
