+++ src/test/java/verify/LottoCountTest.java
@@ -1,9 +1,10 @@
 package verify;
 
+import static model.LottoCount.ONE_LOTTO_PRICE;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static verify.LottoCount.ONE_LOTTO_PRICE;
 
+import model.LottoCount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
